refactor(scripts): split geom-children manifest from emission CLI
Move the SHELL_MANIFEST ledger and completeness helpers into
gen-geom-children-manifest.ts; keep pure shell/index emission and the
package.json entrypoint in gen-geom-children.ts. Deduplicate component
naming through pascalGeom and drop the unused generateGeomChildren path
(CLI uses defineArtifactGroup only).
